+++ client/components/SearchBar.jsx
