+++ client/components/index.js
@@ -4,10 +4,10 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {default as Navbar} from './navbar'
+export {default as Navbar} from './Navbar'
 export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
-export {default as Home} from './Home.jsx'
-export {default as SearchBar} from './searchbar'
-export {default as Graph} from './graph'
-export {default as SingleGraph} from './SingleGraph.jsx'
+export {default as Home} from './Home'
+export {default as SearchBar} from './SearchBar'
+export {default as Graph} from './Graph'
+export {default as SingleGraph} from './SingleGraph'
